requirements.txt
requests
beautifulsoup4
python-whois
tqdm 